greetings class with GetReq
