+++ src/main/java/com/example/app/DemoApplication.java
@@ -1,0 +1,17 @@
+package com.example.app;
+
+import com.example.controllers.myController;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackageClasses = myController.class)
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+
+}
