--- src/main/java/com/example/demo/myController.java
@@ -1,19 +1,0 @@
-package com.example.demo;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.lang.reflect.Method;
-
-@RestController
-public class myController {
-
-    @GetMapping(value="/")
-    public String test(){
-        return "From My Controller";
-    }
-
-
-}
