+++ src/main/java/com/example/services/Greetings.java
@@ -1,0 +1,19 @@
+package com.example.services;
+
+public class Greetings {
+    private final long id;
+    private final String content;
+
+    public Greetings(long id,String content) {
+        this.id = id;
+        this.content = content;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
